Resolve "Add test for in-memory"

See merge request objectbox/objectbox-java!127

+++ tests/objectbox-java-test/src/test/java/io/objectbox/AbstractObjectBoxTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2023 ObjectBox Ltd. All rights reserved.
+ * Copyright 2017-2024 ObjectBox Ltd. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,16 +16,8 @@
 
 package io.objectbox;
 
-import io.objectbox.ModelBuilder.EntityBuilder;
-import io.objectbox.ModelBuilder.PropertyBuilder;
-import io.objectbox.annotation.IndexType;
-import io.objectbox.config.DebugFlags;
-import io.objectbox.model.PropertyFlags;
-import io.objectbox.model.PropertyType;
 import org.junit.After;
 import org.junit.Before;
-
-import javax.annotation.Nullable;
 
 import java.io.File;
 import java.io.IOException;
@@ -37,10 +29,21 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Random;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Stream;
+
+import javax.annotation.Nullable;
+
+import io.objectbox.ModelBuilder.EntityBuilder;
+import io.objectbox.ModelBuilder.PropertyBuilder;
+import io.objectbox.annotation.IndexType;
+import io.objectbox.config.DebugFlags;
+import io.objectbox.model.PropertyFlags;
+import io.objectbox.model.PropertyType;
+
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -52,6 +55,11 @@ public abstract class AbstractObjectBoxTest {
      * Turns on additional log output, including logging of transactions or query parameters.
      */
     protected static final boolean DEBUG_LOG = false;
+
+    /**
+     * If instead of files the database should be in memory.
+     */
+    protected static final boolean IN_MEMORY = Objects.equals(System.getProperty("obx.inMemory"), "true");
     private static boolean printedVersionsOnce;
 
     protected File boxStoreDir;
@@ -92,6 +100,7 @@ public abstract class AbstractObjectBoxTest {
             System.out.println("ObjectBox Java version: " + BoxStore.getVersion());
             System.out.println("ObjectBox Core version: " + BoxStore.getVersionNative());
             System.out.println("First DB dir: " + boxStoreDir);
+            System.out.println("IN_MEMORY=" + IN_MEMORY);
             System.out.println("java.version=" + System.getProperty("java.version"));
             System.out.println("file.encoding=" + System.getProperty("file.encoding"));
             System.out.println("sun.jnu.encoding=" + System.getProperty("sun.jnu.encoding"));
@@ -105,11 +114,19 @@ public abstract class AbstractObjectBoxTest {
      * This works with Android without needing any context.
      */
     protected File prepareTempDir(String prefix) throws IOException {
-        File tempFile = File.createTempFile(prefix, "");
-        if (!tempFile.delete()) {
-            throw new IOException("Could not prep temp dir; file delete failed for " + tempFile.getAbsolutePath());
+        if (IN_MEMORY) {
+            // Instead of random temp directory, use random suffix for each test to avoid re-using existing database
+            // from other tests in case clean-up fails.
+            // Note: tearDown code will still work as the directory does not exist.
+            String randomPart = Long.toUnsignedString(random.nextLong());
+            return new File(BoxStore.IN_MEMORY_PREFIX + prefix + randomPart);
+        } else {
+            File tempFile = File.createTempFile(prefix, "");
+            if (!tempFile.delete()) {
+                throw new IOException("Could not prep temp dir; file delete failed for " + tempFile.getAbsolutePath());
+            }
+            return tempFile;
         }
-        return tempFile;
     }
 
     protected BoxStore createBoxStore() {
@@ -160,10 +177,13 @@ public abstract class AbstractObjectBoxTest {
                 logError("Could not clean up test", e);
             }
         }
-        deleteAllFiles(boxStoreDir);
+        cleanUpAllFiles(boxStoreDir);
     }
 
-    protected void deleteAllFiles(@Nullable File boxStoreDir) {
+    /**
+     * Manually clean up any leftover files to prevent interference with other tests.
+     */
+    protected void cleanUpAllFiles(@Nullable File boxStoreDir) {
         if (boxStoreDir != null && boxStoreDir.exists()) {
             try (Stream<Path> stream = Files.walk(boxStoreDir.toPath())) {
                 stream.sorted(Comparator.reverseOrder())
@@ -216,6 +236,13 @@ public abstract class AbstractObjectBoxTest {
         modelBuilder.lastEntityId(lastEntityId, lastEntityUid);
         modelBuilder.lastIndexId(lastIndexId, lastIndexUid);
         return modelBuilder.build();
+    }
+
+    /**
+     * When not using the {@link #store} of this to create a builder with the default test model.
+     */
+    protected BoxStoreBuilder createBuilderWithTestModel() {
+        return new BoxStoreBuilder(createTestModel(null));
     }
 
     private void addTestEntity(ModelBuilder modelBuilder, @Nullable IndexType simpleStringIndexType) {
